Доработки CD

+++ update_and_rebuild.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 
-# Pull the latest image based on the IMAGE_TAG environment variable
-docker pull artfaal/wisdomwave:${IMAGE_TAG:-other-latest}
+# Ensure that the IMAGE_TAG variable is set
+if [[ -z "$IMAGE_TAG" ]]; then
+    echo "ERROR: The IMAGE_TAG variable is not set."
+    exit 1
+fi
 
-# Use docker-compose to update your running containers
+# Pull the latest image version
+docker pull artfaal/wisdomwave:$IMAGE_TAG
+
+export IMAGE_TAG
+# Navigate to the directory containing docker-compose file
+# Note: This assumes that the docker-compose file is in the same directory from which this script is run.
 docker-compose down
 docker-compose up -d
